+++ Cheetah/Tests/CheetahWrapper.py
@@ -596,5 +596,3 @@ def main():
 
 if __name__ == '__main__':
     main()
-
-# vim: sw=4 ts=4 expandtab