+++ docs/news.rst
@@ -4,6 +4,12 @@ News
 Development (master)
 --------------------
 
+Tests:
+
+  - Removed ``unittest.main()`` calls from tests:
+    ``python -m unittest discover -t Cheetah -s Cheetah/Tests -p '[A-Z]*.py'``
+    does it.
+
 3.2.2 (2019-04-29)
 ------------------